updated properties to ease comp speed
