+++ environment/MarketEnv.py
@@ -1,26 +1,37 @@
+from re import A
 import numpy as np
-import itertools
-
-# np.random.bit_generator = np.random._bit_generator
 
 class MarketEnv():
+    def __init__(self, 
+        action_size, 
+        max_price = None, 
+        max_demand = 50, 
+        
+        n_agents = 1, 
+        max_inventory = 2500, 
+        max_belief = 1, 
+        beta0 = 25,
+        beta1 = -1.1, 
+        beta2 = -2, 
+        a = 0.1):
 
-    def __init__(self, action_size, max_price = None, max_demand = 50, demand_slope = 5, n_agents = 1, max_inventory = 2500, max_belief = 1):
-        # MARL parameters
         if max_price is None:
             max_price = action_size # increment of 1
+        
+        self.beta0 = beta0
+        self.beta1 = beta1
+        self.beta2 = beta2
+        self.a = a
+        self.demand_slope = self.beta0 +  self.beta1 + self.beta2
 
-        self.visit_counter = dict()
         self.n_agents = n_agents
-
-        self.max_inventory = max_inventory
         self.state_env_dim = n_agents + 1
         
+        self.max_inventory = max_inventory
         self.inventory = self.max_inventory
         self.inventory_space_single = np.arange(0, max_inventory+1)
 
         self.action_space = np.arange(0, action_size) * (max_price / action_size)
-        
         self.action_space_joint = np.arange(0, action_size * n_agents)
         self.action_size = action_size
         self.action_env_dim = len(self.action_space)
@@ -29,35 +40,28 @@ class MarketEnv():
         self.current_state = np.array([self.inventory]*self.n_agents + [random_price])
 
         self.max_demand = max_demand
-        self.demand_slope = demand_slope
 
         # State space
         comb_arg = [self.inventory_space_single] * self.n_agents + [self.action_space] # includes reference price
         self.state_space = np.array(np.meshgrid(*comb_arg)).T.reshape(-1, self.n_agents + 1)  #.T.reshape(-1, self.n_agents)
-        
         self.state_space_size = len(self.state_space)
 
         self.max_belief = max_belief
-    
-    def seed(self, seed):
-        return None
-    
-    # state is defined as inventories of each agent + 
+
     def reset(self):
         random_ref_price = np.random.uniform(0, self.action_size)
         self.inventory = self.max_inventory
         return np.array([self.inventory] * self.n_agents + [random_ref_price])
     
-    def get_random_price(self):
-        return np.random.uniform(0, self.action_size)
-        
-    # .step(action)
-    # Takes price index as step argument
-    def step(self, action_index, n_agents = 1):
+    def step(self, action_index):
         
         # previous_ref_price = self.current_state[-1]
         set_price = self.action_space[action_index] # needs edit
-        demand_lambda = self.max_demand - self.demand_slope*set_price
+
+        ref_price = self.current_state[-1]
+        y_intercept = -self.beta2 * ref_price
+
+        demand_lambda = y_intercept - self.demand_slope*set_price
         clipped_lambda = np.max([demand_lambda, 0])
         demand = np.floor(np.random.poisson(clipped_lambda))        
         
@@ -76,14 +80,18 @@ class MarketEnv():
         
         return next_state, reward, inventory_limit, dict()
     
-    def joint_step(self, action_indices):
+    def joint_step(self, action_indices, n_agent_index = 0):
         
         # previous_ref_price = self.current_state[-1]
         action_values = action_indices # special case
         # selected_joint_act_index = int(np.max(action_values)) # refernce price: **minimum or average or any other function**, current or previous time
         # set_price = self.action_space[selected_joint_act_index]
         set_price = self.current_state[-1]
-        demand_lambda = self.max_demand - self.demand_slope*set_price
+        y_intercept = -self.beta2 * set_price
+
+        demand_lambda = 0
+        for a in range(self.n_agents):
+            demand_lambda += y_intercept - self.demand_slope*self.action_space[int(action_indices[a])]
 
         clipped_lambda = np.max([demand_lambda, 0])
         demand = np.floor(np.random.poisson(clipped_lambda))        
@@ -121,12 +129,12 @@ class MarketEnv():
         next_state = self.current_state
         
         return next_state, rewards, [inventory_limit] * self.n_agents, dict()
-
+    
     def map_belief(self, x):
-            x_intercept = self.max_demand / self.demand_slope
-            belief_slope = self.max_belief / x_intercept
-            return self.max_belief - belief_slope*x
-
+        x_intercept = self.max_demand / self.demand_slope
+        belief_slope = self.max_belief / x_intercept
+        return self.max_belief - belief_slope*x
+    
     def auction_system(self, agent_actions, demand):
 
         # assume indices is price value
@@ -144,7 +152,3 @@ class MarketEnv():
             auction_win_agent[vi] += 1
         
         return auction_win_agent
-    
-    def price_translate_from_index(self, i):
-        return float(i)
-
